remove messaging cast on the kernel debug adapter
